Use wp standard structure for theme translations
Before the var `translatePath` was not used and a folder with the name of the .pot file was created, containing the .pot file. As its the Wordpress approach to use a `./languages` folder, I would recommend to add `translatePath` again, change it to follow the naming convention of the other variables (~SRC, ~Destination, ~File) and change the script to use `translatePath` again.